Pushing crash handling on app domain into CoreRuntime
